Conflicts:
	SnifferPy/.idea/workspace.xml
	SnifferPy/src/virsualize/MethaneBuilder.py
	bridge/cfg/detection.json
	bridge/cfg/hypothesis.json
	bridge/output/msg.bin

+++ SnifferCxx/SnifferCxx/model/Coordinate.h
@@ -18,7 +18,7 @@
 namespace Model {
 	class WindVector;
 
-	typedef long coord_item_t;
+	typedef int coord_item_t;
 	typedef std::array<coord_item_t, 3> coord_t;
 
 	class Coordinate : public coord_t {
